Introduce pokemon._suppressed_ability
This new attribute is used to temporarily store an ability when it gets
suppressed. Previously, pokemon._ability was used for this purpose, but it
cannot be used for pokemon whose ability has changed (due to trace, mummy, etc)
since the original ability would be lost.
BattlePokemon._ability is now guaranteed not to change.