Skip klutz
Klutz appears in approximately 0.0005% of battles, and the AI will still be able
to trick assaultvest etc. since it gets its choices from the server. The
overhead of implementing checks for Klutz/item-suppression on every item access
is probably too high to be worth it, though it is probably worth experimenting
with later, or if klutz gets higher usage.