BattleClient: Track foe's original item
This is needed for more accurate deduction of foes' movesets. For
example, a foe could lose its item then use Trick to gain a new item on
the same turn. The original item may be useful for determining the rest
of the set (for example, a Flame Orb conkeldurr always has Guts).