BattleClient tests: make tests more realistic
The next commit will begin inferring and setting moves/item/ability of pokemon
as they're sent out. Tests in this commit are modified to more accurately
reflect a real battle, e.g. pokemon that cannot get a move are replaced with
ones that do. When checking for hiddenpower types, it is no longer assumed that
hiddenpower is the only move in the pokemon's moveset.