Remove effects.Trapper
The trapper/trappee circular relationship was more complicated than it
needed to be. Removing trapper and letting the effects.Trapped and
effects.PartialTrap be independent of the trapper is far simpler. The
BattleEngine now just calls foe.remove_trap_effects when a pokemon
switches out or faints (instead of relying on the Trapper effect to
handle this using on_end).