BattleClient: fix for get_pokemon_from_msg
Require that side.remaining_pokemon > 1 when assuming that a message
refers to my zoroark. Checking the first pokemon in the request message
on a switch after a fainted zoroark is insufficient on the last turn of
the game (e.g. final switch-in faints from hazards), because a new
request object is not sent (so the current request is still the old
one).

In any case, if there's only one remaining pokemon, we don't need to
override the name with zoroark because it can't be confused for any
other pokemon. (the case where we get a |faint| after zoroark has
fainted from |damage| is handled by the first part of the condition,
because zoroark is still the side.active_pokemon)