BattleClient: fix detection of ally zoroark on a switch
The previous implementation fails to detect the zoroark in this case:

(mega-swampert is currently active)
|switch|p2a: Tornadus|Tornadus-Therian, L75, M|222/222 <--actually zoroark!
|move|p1a: Gyarados|Dragon Tail|p2a: Tornadus
|replace|p2a: Zoroark|Zoroark, L78, M|222/222
|-end|p2a: Zoroark|Illusion
|-damage|p2a: Zoroark|164/222
|drag|p2a: Swampert|Swampert-Mega, L77, F|75/281

because the request object shows swampert-mega as the active pokemon at
the end of the turn.

It now relies instead upon zoroark's max_hp being unique (222), since it
is shown to its owner in the |switch| message. If this becomes
non-unique, I'll have to think of something else.