Combine pokemon.moveset + pokemon.pp -> pokemon.moves
moveset was a list, while pp was a dict with the keys redundantly being
the elements of moveset. This commit removes pokemon.moveset, and
renames pokemon.pp -> pokemon.moves. All users of BattlePokemon are
changed to use pokemon.moves to access/iterate over the moves, and a .pp
property is added to BattlePokemon to alias .moves for pp lookup.