BattleClient: fix ally zoroark detection by remembering switch choices
There should only be at most two ally switches in a turn: an intentional
switch and a drag. We can fix the corner cases where both of these occur
and we don't detect the zoroark in-between, and/or when the switch
message doesn't provide enough information to tell which pokemon was
sent out, e.g.:

|switch|p1a: Heliolisk|Heliolisk, L77, F|222/222

We now determine the |switch| pokemon based on the intended switch sent
to the server, and |drag| based on what ends up as the active pokemon
for the next turn as seen in the request object.