Separate creation and usage of rbstats objects
Previously each module would read or create its own RandbatsStatistics
object, which was inefficient. Now instead of automatically creating one
if it doesn't exist, we error out on the missing rbstats.pkl file and
tell the user to create one with `BillsPC.py mine`. Then we read it only
once, in the new file showdowndata/rbstats.py, and use the same object
everywhere.

This speeds up tests by a factor of about 2, and results in less memory
usage overall.