AI: implement and test matrixtree
matrixtree.py includes BreakpointBattle and various types of MatrixNode,
which represent branching points in a game caused by player
decisions. Later versions will represent chance nodes as well.

BreakpointBattle, along with the Breakpoint exception subclasses, allow
simulation of a battle until the next branch point. Whenever a branching
point occurs, the battle is copied once for each possible outcome, which
is represented as another MatrixNode. Each MatrixNode can then continue
the simulation to the next branch point (or end of battle), or use an
evaluation function to determine its value in that state.